Mouse events are properly propagated
We can track the state of the mouse over several widgets. Clicks are
sent with correct coordinates (on X11 at least, unsure about win32) and
we can start a click, and detect inside each widget if the click as
succesfull (when we released the mouse button - it was inside the
widgets).

Mouse events sent to widgets have another state: local (the corrdinates
are relative inside the widget) or global (coordinates are on the
window). Hover events are always local. Widgets get notified regarding
release events - even if the click did not originate inside them.