Started working on text field
The main issue is to be able to send events from th ewidget, up to the
platform - and ask to redraw a window.

Without this - the background color of the text field does not change
(you will need to move the mouse on the main window, to pump the event
system).

Reduced includes in several headers - which caused most of this patch.