Scroll content of the text inside a textfield.cpp
If the cursor is ourside the widget - the widget will display part of
the text, keeping the cursor at its most rightful location.

Scrolling left back, will move the displyed text back keeping the cursor
at the left.

Painting is still borked, as we don't have text clipping yet. Possible
solutions: display partial text, depending on the width of the widget -
or do it at the painting function.