Creating theme entries for TextField
Background of the textfield is no longer hardcoded (its kinda wrong on
all themes, but this is a start). I need some work on each of the
themes, and then I will work on text display.

The text display should also be part of the theme, as padding and
selection is dependent on theme.