Widget + Window: return added value
When you add a Widget to a Window or another Widget - you get the exact
widget you added - with the same type. Previously the library returned a
generic *Widget, but now it will return the same type you added (done
via a templated function, which I should optimize soon).

As a bonus - fixed an out of bound issue in bitmap's pixel method.