Widget: Event propagation
If a sub-widget handles an event - the parent window should not handle
it locally. For example: a spinbox, where the "up button" is pressed,
the local widgets `on_click()` should not be called (as its been
handeled by a sub widget).

This lays the ground for this event, and its not properly implemented
correctly.

As a bonus, moved implementation of some methods of `Widget` to the
source, instead of header file. In case I want to change the
implementation.

The reason for this change:
When you click the up/down button in a spinbox, the event handler of the
textarea should not get called. Now it thinks a click has been done at
the end of the widget - and improperly tries to change the cursor
position.