RadioButtonGroup: draw background
It seems we did not fill the background of the widget with the
corresponding theme. Added code to draw the default widget background to
the base class and "boom"! RadioButtonGroup looks cool.

However... some widgets already draw they own background. And some have
sub-widgets. So - no need to "always clear up background". New flag in
widget - and everything is as before.

COOL.