RadioButtonGroup/Checkbox: Display radio buttons as circle
Theme now renders (ugly ass...) radio buttons. Checkboxes have  a new
property - that marks how they should be drawed. RadioButtonGroup will
set this ro be a circle (so in theory you could have a rounded checkbox,
but that is lame, don't).

Themeing is ugly, but this is good enough for now.