Combobox+Widget: display selected background on focus
New feature: you can set the default background color of a widget from
the widget itself (as it was possible with a whole window).

This is used to set the background color of the combox box using when
its selected. This will fall down when you want a more styled selection
background (instead of a solid color) - but this is good enough for now.