Combobox: focus list on first view
When we open the listview - do focus it, so it gains the arrow up/down
suport. This did not work initially - due to a bug in `PlatformWindow` -
when we focused a widget, we did not invalidate the window. So the popup
would be drawed at a later stage, randomly. Most of this PR is just to
make sure that the focused widget is visible.