MainWindow: remove widget collection
In most UI tookits (GTK, Qt) a window is a kind of widget. I did not
want to go that path, but since both windows and widgets share a lot of
code (ideas, and implementation) I decided to take the shared code into
a 3rd class.

That was a mistake. Instead - the best solution was to use a central
widget. Now the WidgetCollection is used only by Widget (and can
eventually remove that code).