HorizontalLayout: tested and implemented
Widgets now have a layout. All widgets added to the widgets are added to
that layout.

As a test - I hacked the main app, to show 3 widgets in a row. The first
"relayout" is not ideal. It seems that on X11, at least, I am abusing
how the system works and pushing an expose too soon. This is a fight for
another day.