Layout: auto layout support
If you have an horizontal layout, the size hint for this layout will be:

 - for horizontal size: computed by all widgets
 - for vertical: if at least a single widget
   requesed "auto", then size will become vertical.
   Otherwise - the height of the tallest sub-widget.

The exact oposite is done for vertical layout.

Example 1:
 - widget 1, has size hint (0,0)
 - widget 2, has size hint (0,0)
 - widget 3, has size hint (0,1)

This leads to this layout to be possible.
```
|111|222|333|
|111|222|   |
|111|222|   |
```

problem:
If widget 3 request size bigger than the automatically computed size -
it is expected the widget to grow. Instead the widget will be truncated
inside that layout. (unless it refuses to be resizedd in `relayout()` -
and in such case it will overflow the layout).

I have no way to fix this problem yet, and it requires more archtiecture
changes - something I am not even sure how to accomplish.

Example 2:
a strip of horizonal buttons. As all buttons have histhint of {0,1} -
they button strip will have a static height of "1".

```
| 1111111 | 2222222 | 333333 |
```