Layouts: widgets can dynamically adjust to window size
Widgets can be on a layout, vertical-or-horizontal - and the size of
such widget will change dynamically according to the window size.

Your widget can request a size hint - if a value (height or width) has
"0" - means that the layout system will adjust the value of your widget
for you. If you specify a number - this is it.

Layouts have paddings (spaces between each element in the layout) and
margin (the space between the edges of the widget and the closest widget
to that edge).

Items have weight - meaning, that each automatically sized widget can be
larger or smaller then the others in that layout. By default - all
widgets will have the same size.

You can also add a spacer in a layout - and this space will not be
occupied with a widget. Use this with a weight for nicer layouts.