Widget: set self cursor, if no subwidget is selected
Previously, if we moved to a text editor, and then to the section of the
window which contained no widget - the beam cursor was not re-set. The
solution: if no widget is under the mouse, set "my cursor".

Due to architecture problems, I need to pass "*this" to the function. We
should get rid of widget collection and this issue will get resolved.

We also need to set window of sub widgets. This breaks for example radio
button groups. Or the scroll bar (the buttons have no window).