TabHeader/Theme: refactor the display of tabs
Drawing a single tab is done in a function, to simplify the code.

Size hinting is wrong, as (I think) the widget will need to ask the
theme for size of the widget. Otherwise, the padding will need to be
passed from the widget to the theme.

I will need to re-think about doing other types of tab headers.