Bitmap: more optimization on drawing
This simple patch removes from the calling stack the call to the
operator [], which is slow (it was called 213,000,0000 times in a small
zoom, change tabs and exit). I also saw call for const[] operator with
the same time. I also saw that Bitmap::draw() had weight of 18,000,000.

After this path - operator [] (and const) are gone. Drawing has a weight
of 12,000,000. Its almost usable on full screen here.