all: Widgets should own the text
I created a by - in which all widgets have a string_view as the text.
Said text might be out of sope when drawing (for example, when the
widget was created in a function, and now the string view points to a
non-existing stack variable).

Now the main demo shows many tabs with a generated text (and color!).